feat(chatbot): 补充类型,调整utils维护位置
